+++ MiniSQL/file_manager.hpp
@@ -16,7 +16,7 @@ public:
     const static FileManager shared;
     bool FileExistsAt(const std::string &) const;
     void CreateFileAt(const std::string &) const;
-    uint32_t FileSizeAt(const std::string &) const;
+    uint64_t FileSizeAt(const std::string &) const;
     void RemoveFileAt(const std::string &) const;
     
 };
@@ -32,12 +32,12 @@ void FileManager::CreateFileAt(const std::string &path) const {
     std::ofstream os(path);
 }
 
-uint32_t FileManager::FileSizeAt(const std::string &path) const {
+uint64_t FileManager::FileSizeAt(const std::string &path) const {
     using namespace std;
     fstream fs;
     fs.open(path, ios::in | ios::binary);
     fs.seekg(0, ios::end);
-    return (uint32_t) fs.tellg();
+    return (uint64_t) fs.tellg();
 }
 
 void FileManager::RemoveFileAt(const std::string &path) const {
